% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbFetch,SFSQLResult-method}
\alias{dbFetch,SFSQLResult-method}
\title{Retrieve records from SFSQL query}
\usage{
\S4method{dbFetch}{SFSQLResult}(res, n = -1, ...)
}
\description{
Retrieve records from SFSQL query
}
