% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbSendQuery,SFSQLConnection-method}
\alias{dbSendQuery,SFSQLConnection-method}
\title{Send a query to SFSQL.}
\usage{
\S4method{dbSendQuery}{SFSQLConnection}(conn, statement, ...)
}
\arguments{
\item{conn}{database connection, s created by \code{\link{dbConnect}}}

\item{statement}{OGR SQL, see http://www.gdal.org/ogr_sql.html}

\item{...}{for compatibility with generic}
}
\description{
Send a query to SFSQL.
}
\examples{
afile <- system.file("gpkg/nc.gpkg", package = "sf", mustWork = TRUE)
db <- dbConnect(SFSQL(), afile)
dbSendQuery(db, "SELECT * FROM \"nc.gpkg\" WHERE FID < 1")
}
