% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{SFSQL}
\alias{SFSQL}
\alias{dbConnect,SFSQLDriver-method}
\title{SFSQL}
\usage{
SFSQL()

\S4method{dbConnect}{SFSQLDriver}(drv, DSN = "", readonly = TRUE, ...)
}
\arguments{
\item{drv}{SFSQLDriver created by \code{SFSQL()}}

\item{DSN}{data source name, may be a file, or folder path, database connection string, or URL}

\item{readonly}{open in readonly mode?}
}
\description{
SFSQL driver
https://gdal.org/user/ogr_sql_dialect.html

dbConnect
}
\details{
https://gdal.org/user/ogr_sql_dialect.html
}
\examples{
\dontrun{
## a nothing connection not allowed
## ERR: dbConnect(SFSQL())
afile <- system.file("gpkg/nc.gpkg", package = "sf", mustWork = TRUE)
db <- dbConnect(SFSQL(), afile)
dbSendQuery(db, 'SELECT * FROM "nc.gpkg"')
}
}
