% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{SFSQL_PG}
\alias{SFSQL_PG}
\alias{dbConnect,SFSQL_PGDriver-method}
\title{SFSQL_PG}
\usage{
SFSQL_PG()

\S4method{dbConnect}{SFSQL_PGDriver}(
  drv,
  host = "",
  dbname = "",
  user = "",
  password = "",
  readonly = TRUE,
  ...
)
}
\arguments{
\item{drv}{SFSQL_PGDriver created by \code{SFSQL_PG()}}

\item{host}{database server}

\item{dbname}{database name}

\item{user}{user name if needed}

\item{password}{password if needed}

\item{readonly}{open in readonly mode?}
}
\description{
SFSQL_PG driver, a wrapper for SFSQL to construct the full connection string required by
GDAL from host,dbname,user,password

dbConnect for PostgreSQL via GDAL
}
\details{
https://gdal.org/drivers/vector/pg.html
}
\examples{
\dontrun{
db <- dbConnect(SFSQL_PG(), afile)
dbSendQuery(db,
}
}
