% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\docType{class}
\name{SFSQLConnection-class}
\alias{SFSQLConnection-class}
\alias{show,SFSQLConnection-method}
\title{SFSQL connection class.}
\usage{
\S4method{show}{SFSQLConnection}(object)
}
\description{
SFSQL connection class.
}
\keyword{internal}
