% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\docType{class}
\name{SFSQLDriver-class}
\alias{SFSQLDriver-class}
\alias{dbUnloadDriver,SFSQLDriver-method}
\title{Driver for SFSQL.}
\usage{
\S4method{dbUnloadDriver}{SFSQLDriver}(drv, ...)
}
\description{
Driver for SFSQL.
}
\keyword{internal}
