% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{SFSQL_PG}
\alias{SFSQL_PG}
\alias{dbConnect,SFSQL_PGDriver-method}
\title{SFSQL_PG}
\usage{
SFSQL_PG()

\S4method{dbConnect}{SFSQL_PGDriver}(
  drv,
  host = "",
  dbname = "",
  user = "",
  password = "",
  readonly = TRUE,
  ...
)
}
\arguments{
\item{drv}{SFSQL_PGDriver created by \code{SFSQL_PG()}}

\item{readonly}{open in readonly mode?}

\item{DSN}{data source name, may be a file, or folder path, database connection string, or URL}
}
\description{
SFSQL_PG driver

dbConnect
}
\examples{
\dontrun{
db <- dbConnect(SFSQL_PG(), afile)
dbSendQuery(db,
}
}
